# src/main/resources/public/docs/openapi.yml
openapi: 3.0.0
info:
  version: '1.0.0'
  title: Pulmonis API
...
paths:
...
  /public/schoolings:
    get:
      tags:
        - Public Schooling
      summary: Get all public schoolings for visitor
      description: Get all schoolings for visitors which are published. No authentication needed.
      parameters:
        - in: header
          name: find-latest
          required: false
          schema:
            type: boolean
            default: "false"
          description: Flag to find 2 schoolings.
        - in: header
          name: schooling-id
          required: false
          schema:
            type: number
            format: int 64
            example: 24
          description: Schooling ID. Header for excluding currently selected schooling from recommended schoolings. If provided, it will return 3 recomended schoolings.
        - in: path
          name: page
          schema:
            type: integer
            default: 0
            description: Parameter for specifing page of schoolings to be retrieved. If not provided, page=0 will be returned. Default size of page is 16 and elements sorted by update time.
            example: 3
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchoolingListingResponse"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        '500':
          description: Server Internal Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
  /public/schooling/{schoolingId}:
    get:
      tags:
        - Public Schooling
      summary: Get schooling by id
      parameters:
        - in: path
          name: schoolingId
          required: true
          schema:
            type: number
            format: int64
            minimum: 1
          description: The schooling ID
      description: Get schooling for visitors which is published. No authentication needed.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchoolingResponse"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        '500':
          description: Server Internal Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
...
  /schooling/{schoolingId}:
    get:
      tags:
        - Schooling
      summary: Specific schooling by schooling's id
      description: Return doctor's schooling by schooling's id. User ID is retrieved from JWT token.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchoolingResponse"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        '500':
          description: Server Internal Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
...
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    SchoolingResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/SchoolingData"
        errors:
          $ref: "#/components/schemas/EmptyError"
        status:
          $ref: "#/components/schemas/SuccessfulStatus"
    SchoolingListingResponse:
      type: object
      properties:
        data:
          properties:
            totalItems:
              type: integer
              example: 1
              description: Total items returned.
            schoolings:
              type: array
              items:
                $ref: "#/components/schemas/SchoolingData"
            totalPages:
              type: integer
              example: 1
              description: Total pages.
            currentPage:
              type: integer
              example: 0
              description: Currently selected page.
        errors:
          $ref: "#/components/schemas/EmptyError"
        status:
          $ref: "#/components/schemas/SuccessfulStatus"
...
    SchoolingData:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Schooling ID.
          example: 7
        title:
          type: string
          description: Title of schooling.
          example: Ipsum Lorem
        description:
          type: string
          description: Description of schooling.
          example: Lorem ipsum
        address:
          type: string
          description: Address where schooling takes place.
          example: Lorem
        city:
          type: string
          description: City where schooling takes place.
          example: Ipsum
        category:
          type: string
          description: Category of schooling.
          example: Lorem ipsum
        eventDateTime:
          type: date-time
          description: Schooling event date time in format of yyyy-mm-ddThh:mm:ss
          example: 2015-12-03T10:15:35
        free:
          type: boolean
          description: Whether the schooling is free of charge or not.
          example: false
        price:
          type: double
          description: The price of schooling if schooling not free of charge. Otherwise price defaults to 0.0.
          example: 420.69
        registrationDeadline:
          type: date-time
          description: Schooling event deadline in format of yyyy-mm-ddThh:mm:ss
          example: 2015-12-03T10:15:35
        published:
          type: boolean
          description: Whether the schooling is published or not.
          example: true
...
    EmptyError:
      type: array
      nullable: true
      description: List of errors returned by API.
      example: null
    SuccessfulStatus:
      type: integer
      example: 200
      default: 200
      description: Successful response from API.
    UnsuccessfulStatus:
      type: integer
      example: 400
      description: Successful status response from API. Can vary from 400 to 500.
    UnsuccessfulResponse:
      type: object
      properties:
        data:
          type: object
          nullable: true
          example: null
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ApiError"
        status:
          $ref: "#/components/schemas/UnsuccessfulStatus"
    ApiError:
      type: object
      properties:
        message:
          type: string
          description: Error message returned from API.
          nullable: true
          example: "string"
        code:
          type: string
          description: Error code returned from API.
          example: "string"
security:
  - bearerAuth: [ ]